Completed Save image and diff b/w read & unread
