+++ lib/app/modules/reader/widgets/reader_modes/single_vertical.dart
@@ -4,70 +4,84 @@ import 'package:flutter/services.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:get/get.dart';
 
-import '../../../../generated/locales.g.dart';
-import '../../../data/enums/auth_type.dart';
-import '../../../widgets/emoticons.dart';
-import '../controllers/reader_controller.dart';
+import '../../../../../generated/locales.g.dart';
+import '../../../../data/enums/auth_type.dart';
+import '../../../../widgets/emoticons.dart';
+import '../../controllers/reader_controller.dart';
+import '../reader_page_bottom_sheet.dart';
 
-class ContinuousHorizontalRTL extends StatelessWidget {
-  ContinuousHorizontalRTL.asFunction({
+class SingleVertical extends StatelessWidget {
+  SingleVertical.asFunction({
     Key? key,
     required this.controller,
   }) : super(key: key);
 
   final ReaderController controller;
 
-  final ScrollController scrollController = ScrollController();
+  final PageController pageController = PageController();
+
   @override
   Widget build(BuildContext context) {
     return Shortcuts(
       shortcuts: {
-        LogicalKeySet(LogicalKeyboardKey.arrowRight): PreviousScroll(),
-        LogicalKeySet(LogicalKeyboardKey.arrowLeft): NextScroll(),
+        LogicalKeySet(LogicalKeyboardKey.arrowUp): PreviousScroll(),
+        LogicalKeySet(LogicalKeyboardKey.arrowDown): NextScroll(),
       },
       child: Actions(
         actions: {
           PreviousScroll: CallbackAction<PreviousScroll>(
-            onInvoke: (intent) => scrollController.animateTo(
-                scrollController.offset - 300,
+            onInvoke: (intent) => pageController.animateToPage(
+                (pageController.page!).toInt() - 1,
                 duration: Duration(milliseconds: 500),
                 curve: Curves.ease),
           ),
           NextScroll: CallbackAction<NextScroll>(
-            onInvoke: (intent) => scrollController.animateTo(
-                scrollController.offset + 300,
+            onInvoke: (intent) => pageController.animateToPage(
+                (pageController.page!).toInt() + 1,
                 duration: Duration(milliseconds: 500),
                 curve: Curves.ease),
           ),
         },
         child: Focus(
           autofocus: true,
-          child: ListView.builder(
-            scrollDirection: Axis.horizontal,
-            reverse: true,
-            controller: scrollController,
+          child: PageView.builder(
             itemCount: controller.chapter.pageCount,
+            controller: pageController,
+            scrollDirection: Axis.vertical,
             itemBuilder: (context, index) {
               if (index == (controller.chapter.pageCount! - 1)) {
                 controller.markAsRead();
               }
-              return Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                child: CachedNetworkImage(
-                  fit: BoxFit.fitHeight,
-                  httpHeaders: controller.localStorageService.baseAuthType ==
-                          AuthType.basic
+              Map<String, String>? headers =
+                  controller.localStorageService.baseAuthType == AuthType.basic
                       ? {
                           "Authorization":
                               controller.localStorageService.basicAuth,
                         }
-                      : null,
+                      : null;
+              return GestureDetector(
+                onSecondaryTap: () {
+                  readerPageBottomSheet(
+                    index: index,
+                    controller: controller,
+                    headers: headers,
+                  );
+                },
+                onLongPress: () {
+                  readerPageBottomSheet(
+                    index: index,
+                    controller: controller,
+                    headers: headers,
+                  );
+                },
+                child: CachedNetworkImage(
+                  fit: BoxFit.fitHeight,
+                  httpHeaders: headers,
                   imageUrl: controller.getChapterPage(index),
                   filterQuality: FilterQuality.medium,
                   progressIndicatorBuilder: (context, url, downloadProgress) =>
                       SizedBox(
                     height: context.height,
-                    width: context.width * .7,
                     child: Center(
                       child: CircularProgressIndicator(
                         value: downloadProgress.progress,
