+++ lib/app/modules/reader/widgets/reader_modes/webtoon.dart
@@ -4,66 +4,81 @@ import 'package:flutter/services.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:get/get.dart';
 
-import '../../../../generated/locales.g.dart';
-import '../../../data/enums/auth_type.dart';
-import '../../../widgets/emoticons.dart';
-import '../controllers/reader_controller.dart';
+import '../../../../../generated/locales.g.dart';
+import '../../../../data/enums/auth_type.dart';
+import '../../../../widgets/emoticons.dart';
+import '../../controllers/reader_controller.dart';
+import '../reader_page_bottom_sheet.dart';
 
-class SingleHorizontalLTR extends StatelessWidget {
-  SingleHorizontalLTR.asFunction({
+class Webtoon extends StatelessWidget {
+  Webtoon.asFunction({
     Key? key,
     required this.controller,
   }) : super(key: key);
 
   final ReaderController controller;
-  final PageController pageController = PageController();
-
+  final ScrollController scrollController = ScrollController();
   @override
   Widget build(BuildContext context) {
     return Shortcuts(
       shortcuts: {
-        LogicalKeySet(LogicalKeyboardKey.arrowLeft): PreviousScroll(),
-        LogicalKeySet(LogicalKeyboardKey.arrowRight): NextScroll(),
+        LogicalKeySet(LogicalKeyboardKey.arrowUp): PreviousScroll(),
+        LogicalKeySet(LogicalKeyboardKey.arrowDown): NextScroll(),
       },
       child: Actions(
         actions: {
           PreviousScroll: CallbackAction<PreviousScroll>(
-            onInvoke: (intent) => pageController.animateToPage(
-                (pageController.page!).toInt() - 1,
+            onInvoke: (intent) => scrollController.animateTo(
+                scrollController.offset - 300,
                 duration: Duration(milliseconds: 500),
                 curve: Curves.ease),
           ),
           NextScroll: CallbackAction<NextScroll>(
-            onInvoke: (intent) => pageController.animateToPage(
-                (pageController.page!).toInt() + 1,
+            onInvoke: (intent) => scrollController.animateTo(
+                scrollController.offset + 300,
                 duration: Duration(milliseconds: 500),
                 curve: Curves.ease),
           ),
         },
         child: Focus(
           autofocus: true,
-          child: PageView.builder(
-            controller: pageController,
+          child: ListView.builder(
+            controller: scrollController,
             itemCount: controller.chapter.pageCount,
             itemBuilder: (context, index) {
               if (index == (controller.chapter.pageCount! - 1)) {
                 controller.markAsRead();
               }
-              return Center(
-                child: CachedNetworkImage(
-                  fit: BoxFit.fitHeight,
-                  httpHeaders: controller.localStorageService.baseAuthType ==
-                          AuthType.basic
+              Map<String, String>? headers =
+                  controller.localStorageService.baseAuthType == AuthType.basic
                       ? {
                           "Authorization":
                               controller.localStorageService.basicAuth,
                         }
-                      : null,
+                      : null;
+              return GestureDetector(
+                onSecondaryTap: () {
+                  readerPageBottomSheet(
+                    index: index,
+                    controller: controller,
+                    headers: headers,
+                  );
+                },
+                onLongPress: () {
+                  readerPageBottomSheet(
+                    index: index,
+                    controller: controller,
+                    headers: headers,
+                  );
+                },
+                child: CachedNetworkImage(
+                  fit: BoxFit.fitWidth,
                   imageUrl: controller.getChapterPage(index),
+                  httpHeaders: headers,
                   filterQuality: FilterQuality.medium,
                   progressIndicatorBuilder: (context, url, downloadProgress) =>
                       SizedBox(
-                    height: context.height,
+                    height: context.height * .7,
                     child: Center(
                       child: CircularProgressIndicator(
                         value: downloadProgress.progress,
