+++ pubspec.yaml
@@ -12,12 +12,15 @@ dependencies:
   file_picker: ^4.3.0
   flutter:
     sdk: flutter
+  flutter_cache_manager: ^3.3.0
   get: 4.6.1
   get_storage: ^2.0.3
   grouped_list: ^4.1.0
   infinite_scroll_pagination: ^3.1.0
   intl: ^0.17.0
   package_info_plus: ^1.4.0
+  path: ^1.8.1
+  path_provider: ^2.0.9
   readmore: ^2.1.0
   url_launcher: ^6.0.17
